+++ CHANGELOG.md
@@ -3,14 +3,22 @@
 ## [Unreleased]
 
 - Added `add-docs`, `update-docs` commands to `pyrefdev-indexer`.
+- Support [boto3](https://boto3.amazonaws.com/v1/documentation/api/latest/index.html).
+- Support [setuptools](https://setuptools.pypa.io/en/latest/).
+- Support [botocore](https://botocore.amazonaws.com/v1/documentation/api/latest/index.html).
+- Support [charset-normalizer](https://charset-normalizer.readthedocs.io/en/latest/).
+- Support [typing-extensions](https://typing-extensions.readthedocs.io/en/latest/).
+- Support [packaging](https://packaging.pypa.io/en/stable/).
+- Support [s3fs](https://s3fs.readthedocs.io/en/latest/).
+- Support [six](https://six.readthedocs.io/).
 
 ## v2025.4
 
 - Fix a bug in the crawler where relative paths aren't joined correctly.
-- Support [pandas](https://pandas.pydata.org/docs/reference/index.html)
-- Support [urllib3](https://urllib3.readthedocs.io/en/stable/reference/index.html)
-- Support [requests](https://requests.readthedocs.io/en/latest/)
-- Support [python-dateutil](https://dateutil.readthedocs.io/en/stable/)
+- Support [pandas](https://pandas.pydata.org/docs/reference/index.html).
+- Support [urllib3](https://urllib3.readthedocs.io/en/stable/reference/index.html).
+- Support [requests](https://requests.readthedocs.io/en/latest/).
+- Support [python-dateutil](https://dateutil.readthedocs.io/en/stable/).
 
 ## v2025.3